Implemented the staged object rendering and fixed some crash errors - Implementedt the object rendering in the GraphicManager - Fixed a ClockManager problem causing the app to crash - Created the GameObjectsManager and moved the relatives functions from the GraphicManager - Created the ApplicationMain, the main class for the App

+++ PathfindingAlgorithmsProject/PathfindingAlgorithmsProject/ClockManager.cpp
@@ -47,7 +47,7 @@ void ClockManager::NewFrame()
 
 void ClockManager::ManageFramesCap()
 {
-    float elapsedMS = (SDL_GetPerformanceCounter() - _startFrameCounter) / (float)SDL_GetPerformanceFrequency() * 1000.0f;
+    float elapsedMS = (SDL_GetPerformanceCounter() - _startFrameCounter) / SDL_GetPerformanceFrequency() * 1000.0f;
 
     SDL_Delay(floor((1000 / _frameRateCap) - elapsedMS));
 }
